# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Madrasa & Tuitions Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  email: string;
  password: string;
  role?: string;
}

export interface AuthUser {
  id: string;
  name: string;
  email: string;
  role: string;
}

export interface AuthResponse {
  token: string;
  user: AuthUser;
}

export interface DashboardSummary {
  totalStudents: number;
  totalTeachers: number;
  totalAssignments: number;
  totalLessons: number;
  totalClasses: number;
  activeStudents: number;
  pendingAssignments: number;
  ongoingLessons: number;
  attendanceRate: number;
  completionRate: number;
}

export interface ActivityItem {
  id: string;
  type: string;
  message: string;
  /** @nullable */
  messageUr?: string | null;
  timestamp: string;
  actor: string;
  /** @nullable */
  icon?: string | null;
}

export interface Student {
  id: string;
  name: string;
  /** @nullable */
  nameUr?: string | null;
  email: string;
  phone: string;
  classId: string;
  className: string;
  enrollDate: string;
  guardianName: string;
  /** @nullable */
  guardianPhone?: string | null;
  status: string;
  /** @nullable */
  grade: string | null;
  attendance: number;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface StudentInput {
  name: string;
  /** @nullable */
  nameUr?: string | null;
  email: string;
  phone: string;
  classId: string;
  enrollDate: string;
  guardianName: string;
  /** @nullable */
  guardianPhone?: string | null;
  status: string;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface StudentUpdate {
  name?: string;
  /** @nullable */
  nameUr?: string | null;
  email?: string;
  phone?: string;
  classId?: string;
  enrollDate?: string;
  guardianName?: string;
  /** @nullable */
  guardianPhone?: string | null;
  status?: string;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface StudentList {
  data: Student[];
  total: number;
  page: number;
  limit: number;
}

export interface Teacher {
  id: string;
  name: string;
  /** @nullable */
  nameUr?: string | null;
  email: string;
  phone: string;
  subjects: string[];
  classes: string[];
  qualification: string;
  hireDate: string;
  status: string;
  /** @nullable */
  specialization?: string | null;
  /** @nullable */
  experience?: number | null;
}

export interface TeacherInput {
  name: string;
  /** @nullable */
  nameUr?: string | null;
  email: string;
  phone: string;
  subjects: string[];
  classes?: string[];
  qualification: string;
  hireDate: string;
  status: string;
  /** @nullable */
  specialization?: string | null;
  /** @nullable */
  experience?: number | null;
}

export interface TeacherUpdate {
  name?: string;
  /** @nullable */
  nameUr?: string | null;
  email?: string;
  phone?: string;
  subjects?: string[];
  classes?: string[];
  qualification?: string;
  hireDate?: string;
  status?: string;
  /** @nullable */
  specialization?: string | null;
  /** @nullable */
  experience?: number | null;
}

export interface TeacherList {
  data: Teacher[];
  total: number;
  page: number;
  limit: number;
}

export interface Assignment {
  id: string;
  title: string;
  /** @nullable */
  titleUr?: string | null;
  description: string;
  /** @nullable */
  descriptionUr?: string | null;
  dueDate: string;
  assignedTo: string;
  classId: string;
  className?: string;
  status: string;
  createdBy: string;
  createdAt?: string;
  /** @nullable */
  subject?: string | null;
}

export interface AssignmentInput {
  title: string;
  /** @nullable */
  titleUr?: string | null;
  description: string;
  /** @nullable */
  descriptionUr?: string | null;
  dueDate: string;
  assignedTo: string;
  classId: string;
  status: string;
  createdBy: string;
  /** @nullable */
  subject?: string | null;
}

export interface AssignmentUpdate {
  title?: string;
  /** @nullable */
  titleUr?: string | null;
  description?: string;
  /** @nullable */
  descriptionUr?: string | null;
  dueDate?: string;
  assignedTo?: string;
  classId?: string;
  status?: string;
  /** @nullable */
  subject?: string | null;
}

export interface AssignmentList {
  data: Assignment[];
  total: number;
  page: number;
  limit: number;
}

export interface Lesson {
  id: string;
  title: string;
  /** @nullable */
  titleUr?: string | null;
  classId: string;
  className: string;
  teacherId: string;
  teacherName: string;
  schedule: string;
  duration: number;
  status: string;
  students?: string[];
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  progress?: number | null;
}

export interface LessonInput {
  title: string;
  /** @nullable */
  titleUr?: string | null;
  classId: string;
  teacherId: string;
  schedule: string;
  duration: number;
  status: string;
  students?: string[];
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  progress?: number | null;
}

export interface LessonUpdate {
  title?: string;
  /** @nullable */
  titleUr?: string | null;
  classId?: string;
  teacherId?: string;
  schedule?: string;
  duration?: number;
  status?: string;
  students?: string[];
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  progress?: number | null;
}

export interface Class {
  id: string;
  name: string;
  /** @nullable */
  nameUr?: string | null;
  teacherId: string;
  teacherName: string;
  schedule: string;
  studentCount: number;
  /** @nullable */
  subject?: string | null;
}

export type GetStudentsParams = {
search?: string;
status?: string;
classId?: string;
page?: number;
limit?: number;
};

export type GetTeachersParams = {
search?: string;
page?: number;
limit?: number;
};

export type GetAssignmentsParams = {
search?: string;
status?: string;
classId?: string;
page?: number;
limit?: number;
};

export type GetLessonsParams = {
search?: string;
status?: string;
classId?: string;
};