# lib/api-zod/src/generated/types/assignmentInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Madrasa & Tuitions Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignmentInput {
  title: string;
  /** @nullable */
  titleUr?: string | null;
  description: string;
  /** @nullable */
  descriptionUr?: string | null;
  dueDate: string;
  assignedTo: string;
  classId: string;
  status: string;
  createdBy: string;
  /** @nullable */
  subject?: string | null;
}